[0.5.1] RELEASE (#20)
* add menu announcement (#18)

* fix session channel point
