+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "so",
-  "version": "0.5.0",
+  "version": "0.5.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
